blogify_backend/requirements.txt
asgiref==3.7.2
django==4.2.6
sqlparse==0.4.4
tzdata==2023.3
bleach==6.1.0
django-summernote==0.8.20.0
six==1.16.0
webencodings==0.5.1
Pillow==10.0.1
djangorestframework==3.14.0
pytz==2023.3.post1